Moved Buttons in separate folder, fixed CMakefile
